pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harry-potter"
version = "0.1.0"
description = "Terminal-first Harry Potter exploration sandbox set around Hogwarts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.40.0",
  "PyYAML>=6.0.1",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[project.scripts]
harry-potter = "hogwarts_game.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
hogwarts_game = ["content/*.yaml"]
